Split aeq_log.h, remove import of global function (fix segfault)
